+++ src/kustology/bridge.py
@@ -211,6 +211,7 @@ _initialize_bridge()
 
 from Kusto.Language import GlobalState, KustoCode, KustoFacts
 from Kusto.Language.Editor import FormattingOptions, KustoCodeService
+from Kusto.Language.Parsing import TokenParser
 from Kusto.Language.Symbols import (
     ColumnSymbol,
     DatabaseSymbol,
@@ -230,4 +231,5 @@ __all__ = [
     "KustoFacts",
     "ScalarTypes",
     "TableSymbol",
+    "TokenParser",
 ]